+++ cached_path/common.py
@@ -1,0 +1,4 @@
+from os import PathLike
+from typing import Union
+
+PathOrStr = Union[str, PathLike]
