+++ setup.py
@@ -58,5 +58,5 @@ setup(
         ]
     ),
     install_requires=install_requirements,
-    python_requires=">=3.6.1",
+    python_requires=">=3.7.1",
 )
